# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API SIHIOS-TS",
        "description": "API de la plateforme SIHIOS-TS",
        "contact": {
            "name": "Developer",
            "email": "michel.banh@softskills.ci"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
        "/admin/store": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Créer un nouvel administrateur",
                "operationId": "c06e44c49f82eb22ad4d98ae27b3399a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "first_name": {
                                        "type": "string",
                                        "example": "Jean"
                                    },
                                    "last_name": {
                                        "type": "string",
                                        "example": "Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "jean.dupont@example.com"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "0123456789"
                                    },
                                    "job_title": {
                                        "type": "string",
                                        "example": "Administrateur"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Administrateur créé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Erreur de validation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/all-member": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtenir tous les membres",
                "operationId": "51154243e3b2c2cdb2212eeee70560e9",
                "responses": {
                    "200": {
                        "description": "Liste des membres récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/admin-list": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtenir tous les administrateurs",
                "operationId": "9dc50c908cc3236895638910fca59e3f",
                "responses": {
                    "200": {
                        "description": "Liste des administrateurs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/active": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtenir tous les membres actifs",
                "operationId": "84da673d9ee36996bdf136003185a4c1",
                "responses": {
                    "200": {
                        "description": "Liste des membres actifs récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/pending-member": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtenir tous les membres en attente",
                "operationId": "c995850ecf5a81a301dd85b1bb4aa32a",
                "responses": {
                    "200": {
                        "description": "Liste des membres en attente récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/rejected-member": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Obtenir tous les membres rejetés",
                "operationId": "082794abd572ce81ed9f5e6d1ccd76b1",
                "responses": {
                    "200": {
                        "description": "Liste des membres rejetés récupérée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/User"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/approve-member/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Approuver un membre par ID",
                "operationId": "43398e0d771e1bc6c509600f0397efed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du membre à approuver",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Membre approuvé avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Membre non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/reject-member/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Rejeter un membre par ID",
                "operationId": "5b67b498b354133acb2cde29bd90d9b9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du membre à rejeter",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Membre rejeté avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Membre non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
        "/admin/pending-member/{id}": {
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Mettre un membre en attente après approbation",
                "operationId": "384d8675d2303e5a7eecd1e20d85ffc5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID du membre à mettre en attente",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Membre mis en attente avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Membre non trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "Bearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "User": {
                "title": "Utilisateur",
                "required": [
                    "id",
                    "first_name",
                    "last_name",
                    "email",
                    "phone",
                    "job_title"
                ],
                "properties": {
                    "id": {
                        "description": "L'identifiant unique de l'utilisateur",
                        "type": "integer"
                    },
                    "first_name": {
                        "description": "Prénom de l'utilisateur",
                        "type": "string"
                    },
                    "last_name": {
                        "description": "Nom de famille de l'utilisateur",
                        "type": "string"
                    },
                    "email": {
                        "description": "Adresse e-mail de l'utilisateur",
                        "type": "string"
                    },
                    "phone": {
                        "description": "Numéro de téléphone de l'utilisateur",
                        "type": "string"
                    },
                    "job_title": {
                        "description": "Poste de l'utilisateur",
                        "type": "string"
                    },
                    "status": {
                        "description": "Statut de l'utilisateur (ex: actif, en attente)",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "Bearer": {
                "type": "http",
                "description": "Utilisez un jeton JWT pour l'authentification.",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}